fix: relocate lineagegraph and stateaccessquery under state_access
Move folded lineage_graph_modules and state_access_query packages from work/internal/{lineagegraph,stateaccessquery} to work/internal/services/state_access/{lineagegraph,stateaccessquery} per recorded INV destinations. Update imports and inventory documentation.
