+++ pkg/services/work/internal/invocationreturnpolicy/primary_result.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/portpowered/infinite-you/pkg/services/work/internal/lineagegraph"
+	"github.com/portpowered/infinite-you/pkg/services/work/internal/services/state_access/lineagegraph"
 )
 
 const (
