+++ pkg/services/work/internal/invocationreturnpolicy/types.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"errors"
 
-	"github.com/portpowered/infinite-you/pkg/services/work/internal/lineagegraph"
+	"github.com/portpowered/infinite-you/pkg/services/work/internal/services/state_access/lineagegraph"
 )
 
 var (
