+++ pkg/services/work/internal/services/state_access/stateaccessquery/list_test.go
@@ -6,7 +6,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/portpowered/infinite-you/pkg/services/work/internal/stateaccessquery"
+	"github.com/portpowered/infinite-you/pkg/services/work/internal/services/state_access/stateaccessquery"
 )
 
 func TestListRequestPreparationReturnsDetachedValidatedValues(t *testing.T) {
