+++ pkg/services/work/lineage_contract.go
@@ -3,7 +3,7 @@ package work
 import (
 	"encoding/json"
 
-	"github.com/portpowered/infinite-you/pkg/services/work/internal/lineagegraph"
+	"github.com/portpowered/infinite-you/pkg/services/work/internal/services/state_access/lineagegraph"
 )
 
 const (
